Move body generator for structs to its own module
Cut-paste
