+++ serde_derive/src/de/enum_externally.rs
@@ -5,8 +5,9 @@
 //! ```
 
 use crate::de::enum_;
+use crate::de::struct_;
 use crate::de::{
-    deserialize_struct, deserialize_tuple, expr_is_missing, field_i,
+    deserialize_tuple, expr_is_missing, field_i,
     unwrap_to_variant_closure, wrap_deserialize_field_with, wrap_deserialize_with, Parameters,
     StructForm, TupleForm,
 };
@@ -148,7 +149,7 @@ fn deserialize_externally_tagged_variant(
             cattrs,
             TupleForm::ExternallyTagged(variant_ident),
         ),
-        Style::Struct => deserialize_struct(
+        Style::Struct => struct_::deserialize_struct(
             params,
             &variant.fields,
             cattrs,
